Added security and working towards properly starting maps

+++ frontend/src/components/ServerManager.js
@@ -4,6 +4,23 @@ export default class ServerManager {
   constructor() {
     this._nonplayed = [];
     this._played = [];
+    axios.interceptors.request.use((config) => {
+      const token = localStorage.getItem('access_token');
+      if (token) {
+        config.headers['Authorization'] = `Bearer ${token}`;
+      }
+      return config;
+    },
+    (error => {
+      window.location.href = '/login';
+      return Promise.reject(error);
+    }))
+  }
+
+  get_auth_header(token) {
+    return {
+      'Authorization': token
+    };
   }
 
   async updateNonplayed() {
@@ -13,35 +30,51 @@ export default class ServerManager {
         this.nonplayed = res.data;
       })
       .catch((error) => {
-        console.error(error);
+        if(error.response.status === 401) {
+          window.location.href = '/login';
+        } else console.error(error);
       });
   }
 
   async updatePlayed() {
+    let token = localStorage.getItem('access_token')
+    if (token === null) {
+      window.location.href = '/login';
+      return;
+    }
     const path = 'http://localhost:5000/played';
     axios.get(path)
       .then((res) => {
         this.played = res.data;
       })
       .catch((error) => {
-        console.error(error);
+        if(error.response.status === 401) {
+          window.location.href = '/login'
+        } else console.error(error);
       });
   }
 
   async addMap(workshop_url) {
     const path = 'http://localhost:5000/submitmap';
     axios.post(path, {
-      workshop_url: workshop_url
+      data: { workshop_url: workshop_url }
     })
       .then(() => {
         this.updateNonplayed();
       })
       .catch((error) => {
-        console.error(error);
+        if(error.response.status === 401) {
+          window.location.href = '/login'
+        } else console.error(error);
       });
   }
 
   async removeMap(workshop_id) {
+    let token = localStorage.getItem('access_token')
+    if (token === null) {
+      window.location.href = '/login';
+      return;
+    }
     const path = 'http://localhost:5000/removemap';
     axios.delete(path, {
       data: {workshop_id: workshop_id}
@@ -51,7 +84,9 @@ export default class ServerManager {
         this.updatePlayed();
       })
       .catch((error) => {
-        console.error(error);
+        if(error.response.status === 401) {
+          window.location.href = '/login'
+        } else console.error(error);
       });
   }
